Resolve flatWrapperTarget() when deducing types
This is a bugfix. It turns out that populateFlatFields() kept some
flatWrapperTarget() rules unresolved in schema_source. This commit
resolves it using resolveIfWrapper().

To fix this bug, we also had to stop using field_name.empty() as a proxy
for flattenability in gatherFlatComps(). These wrapper types break that
rule. They never have a name and yet their flattenability depends on
their target type.