Move ruleAt() into call-site generation function
This is a temporary change while we migrate all functions into
strongly-typed return values. The conversion glue for converting
strongly-typed parser result and into JsonLoc depends on the return
value type of that parser result. With this commit, we will be able
to determine this return type for flatWrapperTarget() rules like
AliasRule or QuietMatch.

So far, this has not been an issue since those flat wrapper rules
jsonize values that pass through them. We can now keep them unchanged if
we want.