Fix conversion of optional fields
When converting a parsed result from a generated struct to JsonLoc, we
had a bug in how we treated optional fields. Instead of nullopt values
simply being absent in the output JsonLoc, we kept them as ErrorValue.
This is undesirable for OrRule where the parse was actually successful,
and is quite different from how we had been representing optional
fields.