ConcatFlatRule, OrRule, LoopRule produces types
These rules used to produce parsers that returned JsonLoc. But now they
can return the newly generated structs instead.

It's a monster of a commit, since any change to a return value requires
us to change every generator of every possible caller, in how they
propagate values.

Most of the change is a straightforward change of types, including the
generated code and unit tests. The primary exception is the OrRule's
code generator, which has become much more complicated. This is because
the logic for generating code for a flattenable OrRule is now very
different from that for the OrRule that is not-flattenable.

The special-case in parserResultTraits() can now be removed. Some of the
codegen now requires type-specific code for accessing the parsed result
of a component, which is why ParserResultTraits now has a new value()
method.

I am not going to attempt another monster commit like this in a while.