Strongly typed fields for OutputTmpl
The plan is to slowly change the `OutputTmpl` fields structure to use
strongly typed fields. This commit merely introduces new `typed_fields`,
but keeps the new fields unpopulated.

Later commits will populate them with the correct values, slowly migrate
the rest of the codebase to this new `typed_fields`, and finally delete
the old JsonLoc `fields`.