Resolve OutputTmpl child type at compile-time
Previously OutputTmpl made a runtime check to see if the child type is a
(flat) map or not. But now we have the tools to make this check at
compile-time, and generate only the appropriate branches.

This will also help us migrate away from the JsonLoc `fields` of
OutputTmpl to the new strongly-typed fields. This way we will no longer
have to depend on `JsonLoc::Map` types in the generated parser.