Return nullopt from errorValue()
This utility is for generating handy one-liners in ErrorRule code.
However, we are about to stop returning JsonLoc::ErrorValue from those
rules. So we can now return a simpler type that has no semantic meaning.

After this commit, the only place where we still kind of use the return
value from errorValue() is in codegenParserCall().

We choose std::nullopt_t instead of void type since a lot of places we
copy and move the return value from generated parsers. The void type
just make that awkward, and requires a lot of special-casing.

It fits in nicely with the fact that most generated parsers use an
std::optional value. They too use nullopt to indicate a failure or
error. So it makes sense that a function that always returns an error
now has the return type (nullopt_t) to reflect that.