Return nullopt from ErrorRule
This is the last rule type that used to return JsonLoc from
generated code. Now it returns a more specific type, std::nullopt_t. See
the previous commit for why we chose nullopt_t as the return type
instead of void.

This happens to be the only flattenable rule type that doesn't produce a
struct. It has to remain flattenable since it frequently appears as a
branch of a flattenable OrRule. The whole reason that ErrorRule was
introduced was to be used in that context. On the other hand, it doesn't
return any useful parsed result, so it doesn't need to return a struct
value either.