Refactor compilation out of lookaheadRuleIndex
This is one step of moving compilation steps of multi-choice rules out
of the frontend. It splits the old function lookaheadRuleIndex() into
two: parseLookahead() in frontend.cpp and appendLookahead() in
compiler.cpp. The first one produces a RuleExpr, while the second one
consumes it.

No change in functionality, but if we later want to support richer
RuleExpr in lookaheads (e.g. regex), we can do it more easily now.