Separate out or-branch compilation from parsing
This commit uses the RuleBranch type to communicate between the parsing
and compiling steps now, instead of directly modifying OrRule.
parseBranchAction() now returns RuleBranch instead of modifying OrRule
and taking in RulesWithLocs for compilation. This RuleBranch is then
consumed in consumeRuleBranch().

The simpler function for appending branches to an OrRule is now more
complicated because of error-handling. The data structure can represent
the totality of features that we _plan_ to implement in "actions" of
multi-choice rules. But right now, most of those steps are marked
"unimplemented".