Push lookahead compilation into passthroughBranch
passthroughBranch() was taking in an extra lookidx parameter even though
it can already accept it through the RuleBranch parameter. This commit
uses RuleBranch::lookahead for this purpose, in effect, moving branch
compilation into that one function.

As a result of this commit, appendLookahead() is now an internal
function of compiler.cpp, and is not exposed in compiler.h.