Accept RuleStanzas in appendExprRule()
The number of parameters into appendExprRule() became large a while ago.
A while back, we had a todo to pack this all into a single struct. We
resolve this todo here now.

We plan to add the exact same parameters in appendMultiExprRule() too.
This commit will let us add only a single parameter instead of the whole
bundle separately.