Accept RuleStanzas in appendMultiExprRule()
This commit is part of introducing RuleExpr support in multi-choice
rules. It changes the interface between the frontend and the compiler to
use RuleStanzas, so that we can slowly support various stanzas in
multi-choice rules. The new RuleExpr won't be useful unless we support
`lexical:` stanzas (for double-quoted expressions) and `where:` stanzas
(for any expression with named subcomponents).

This commit actually starts calling parseRuleStanzas() for multi-choice
rules. But the still unsupported stanzas are temporarily rejected at the
very start of the compilation process, in appendMultiExprRule().