Generate forward declarations for struct rules
For any generated function of this type:

    std::optional<ParsedFoo> parseFoo(InputDiags& ctx, ssize_t& i);

we do need the return type `ParsedFoo` to also be declared before we use
it in the function declaration above. Before this commit, we had been
getting lucky. We just happened to be generating type definitions before
the functions producing results of that type. This function eliminates
the luck factor using forward declarations.

We will stop getting lucky as we start modifying how multi-match rules
are compiled. In later commits, we will make multi-match rule branches
accept arbitrary RuleExpr targets.

For now, it might be redundant in that it may already have been declared
multiple times in this file, or even fully defined. In future, we can
clean it up by keeping track of declared and defined types. For now,
this is good enough.