Use RuleExprCompiler to compile expression rules
This commit pertains to branches of this form:

  | lookahead -> ident

Instead of directly calling a simple findOrAppendIdent(), this commit
switches to using RuleExprCompiler instead. It has no functional change
for now, but it will later help us generalize that `ident` bit into any
RuleExpr.

For now, this commit keeps the guard that constrains the new logic to
RuleExprIdent branches only. It's an internal preparation only. Over the
next few commits, we will generalize the compiler and the frontend, so
that other kinds of RuleExpr are also supported.