Accept arbitrary RuleExpr in the frontend
This is one step of introducing RuleExpr support in multi-choice rule
branches. This commit changes the frontend only, while later commits
will add support in the compiler.

In this commit, we delete the custom branch-parsing function
parseSingleIdentBranch(), and replace it with a more general
makeRuleExpr(). For now, the compiler still enforces that only
RuleExprIdent is provided, so no enhancements are yet user-visible.