Remove duplicate function ruleExprMakeOutputTmpl()
Before this commit we had two functions which were basically duplicates:

  * ruleExprMakeOutputTmpl()
  * deduceOutputTmpl()

Both took a vector<IdentUsage> as input, and produced a JsonTmpl.

The first one removed duplicate identifiers from a sorted list, while
the second one used all the input. This commit keeps the first function,
but uses the name of the second one (just because it's a shorter name).