Split assignNonMapRuleExpr to two functions
The new functions are assignIfStringExpr and the now public lookupIdent.
They handle RuleExprIdent in a separate function. This is the first step
in fixing a bug in how we handle string, regex, or single-ident rules
with an output template. E.g.

    rule wrapped_keyword: 'func'
    outputs: {
      keyword: 'kw_func'
    }

For rules like this, appendExprRule() returns too quickly when
createNonMapRuleExpr() returned success. This is the correct behavior
_if_ there is no output template. To fix this, this function has to know
if createNonMapRuleExpr() processed a RuleExpr{Squoted,Regex} versus a
RuleExprIdent. The latter will produce a new field for template
replacement. One way to have this information in appendExprRule() is by
having it call two separate functions for ident case vs. the rest.

This is what this commit accomplishes. This commit just splits the
function into two parts. The actual bug is fixed in a later commit.