Export ident list from compileSingleExpr()
The return type of this method is changed from unique_ptr<RuleExpr> to
optional<struct>, where the new struct type has both the old RuleExpr
pointer as well as a vector of identifiers found in the compiled rule.

This will help us fix a bug. Currently, we don't properly handle output
templates for RuleExpr{Ident,Squoted,Regex}. This commit will allow us
to export the list of identifiers used from the point where it is
compiled to the point where we use it in checkUnusedParts().