Separate template code path from non-template
This commit bypasses most of appendExprRule() if no `outputs:` template
is provided for the rule (i.e. if stz.jstmpl.holdsEllipsis() is true).
In that cases, it simply reuses compileSingleExpr() instead of
reimplementing everything here.

This will help us fix a bug where we need special logic just for the
case where a template is provided.

While making the change, we noticed a missing somePatternFailed_ check
in compileSingleExpr().