Fix unused ident check
During compilation of a given rule, we collect two sets of identifiers:
the ones that were used by the rule, and the ones that were exported. We
had confused the two. So when calling checkUnusedParts(), I had
mistakenly used the list of idents exported by the rule. This caused
problems for RuleExprMappedIdent.

Most of this commit renames vectors so we know which is which. A few
places we explicitly call ruleExprCollectInputIdents() to collect the
right set of identifiers to return.