codegen() can now suppress missing optional fields
The output of multi-choice rules used to be littered with empty json
dictionaries `{}`. Most of those can now be automatically removed when
converting outputs to JsonLoc. They cause unnecessary noise in example
tests, and is an additional cognitive load for newcomers.

I'm okay with non-uniform semantics if it makes 90% of use cases saner.

This change only affects codegen(). The next commit will make the change
to eval() too.