Implement mapNestedAppend()
This is a helper function to be used in compiled code. When optional
fields are present, converting from a generated struct type to a JsonLoc
requires being able to add key-value pairs after the non-optional part
of the JsonLoc has already been constructed. This may be arbitrarily
deep, depending on what OutputTmpl has been provided by the user.