Resolve old fixme
This commit is merely a cleanup, even though it was (mis-)tagged as a
fixme. We had previously tagged it as a "fixme" since we were returning
OrRules with -1 targets, which is invalid. However, this invalid rule
will never actually be used in practice. parseOalexSource() already
returns nullopt in the face of any error diagnostics at all, so this is
unlikely to lead to any undefined behavior.