Fix undefined output checks
This bugfix concerns how we check for output fields that were never
defined. The code was written long before RuleExprMappedIdent was
introduced, and it checked only identifiers _used_ in a pattern, not the
ones exported from it. These two sets of idents used to be the same, but
no more.

The check for undefined output fields was previously being done as part
of checkUnusedParts(), which was the wrong place:

  * It was using an ambigously named `patternIdents` parameter, which
    encoded identsUsed, not identsExported.
  * It was considered a soft failure which did not return any
    success/failure status. An undefined output needs to be a harder
    failure.

This commit splits the new function checkUndefinedOutfields() out of the
old checkUnusedParts(). The old function no longer needs the jstmpl
parameter, and the param patternIdents have been appropriately renamed.
The new function can now be called as part of compileSingleExprWithTmpl.

We cannot yet easily unit-test this fix, unfortunately. The same test
cases that would be appropriate here also bumps into a separate,
unrelated bug that is not yet fixed. Here is an example test that is
affected once we have that other bug fixed too:

rule foo: (kw ~ "const")
outputs: { kw }