Fix named pattern handling for output fields
In the past, any `outputs` field that was not locally defined caused a
global rule to be expected for that name. This expectation is now
ill-founded, now that we have named patterns. Resolution of `outputs`
field now goes as follows:

  * Search for locally defined rules (in `where:`).
  * Search for a named pattern, e.g. `id` in `(id ~ "patt")`.
  * Finally, look for global rules.

This commit fixes this problem. mapToRule() no longer calls
`findOrAppendIdent()` if the name has already been mapped to a pattern
using the form in the second bullet item above.

We discovered many other problems with the codegen of these rules (as
opposed to the eval() being fixed here). They are unrelated, and will
have to be fixed much later.

This commit also updates the comment for RuleExprCompiler fields. I
noticed that it doesn't talk about `outputs:` template at all, which
confused me as I was examining mapToRule().