Change exportedIdents to an output struct field
This commit makes no functional change, except for moving the output
parameter to the return value. This will be useful in the next commit,
where we will introduce yet another output field.

There is no longer any need to clear the set of exported idents. We
don't really use that to deduce templates in every branch anymore.