Added field CompiledSingleExpr::exportedIdents
The field is already there in CompiledRuleBranch. This commit doesn't do
anything with the new field, just defines it. Later commits will
populate it with a value and use it.

This change will help us clean up the interface of compileRuleExpr() so
that we don't need a separate output field. Right now it's inconsistent:
some methods use a struct to return everything, while some use output
params. Worse still, identsUsed still need to be collected separately by
every caller. After exportedIdents is properly populated, we can use
this struct to return everything.

The real reason for this change is an upcoming bugfix. OutputTmpl isn't
always receiving a flattenable map like it needs to. Like this rule
currently crashes:

    rule dbl_quoted_greeting: "hello [world]"

We need compileRuleExpr() to return the output type of rule it produces
(e.g. map vs. string). Right now, that's impossible for double-quoted
patterns. This means compileRuleExpr() can no longer return just a
`unique_ptr<Rule>`, it needs a struct. It makes sense to use
`CompiledSingleExpr` instead of defining a new struct.