Enrich return type of compileToFlatStruct()
Previously, it used to return unique_ptr<Rule>, and it took in an output
parameter exportedOutput. It also used to expect its callers to collect
identsUsed manually, instead of just returning it as a SortedIdents
vector.

This commit changes this method to return all of these items into a
packaged struct (CompiledSingleExpr). The call to
ruleExprCollectInputIdents() have also been folded in, so callers no
longer have to remember to do that either. This commit is a step in a
refactoring that will also change the return value of compileRuleExpr().

For motivation, see the commit that added the field
CompiledSingleExpr::exportedIdents.