Enrich return type of compileRuleExpr()
This is analogous to the previous commit.

Previously, it used to return unique_ptr<Rule>, and it took in an output
parameter exportedOutput. It also used to expect its callers to collect
identsUsed manually, instead of just returning it as a sortedIdents
vector.

This commit changes this method to return all of these items into a
packaged struct (CompiledSingleExpr). The call to
ruleExprCollectInputIdents() have also been folded in, so callers no
longer have to remember to do that either. This commit finishes the
refactoring that changed the return type of two methods and added
exportedIdents as a field of CompiledSingleExpr.