Use the new types in compileToFlatRule()
This commit is merely a cleanup, not a change in functionality.
It is meant to make the code more robust against future changes.

Previously, compileToFlatRule() had to guess the output type from
the runtime type of the input RuleExpr, even before compileRuleExpr()
was called. This can be a problem if we later introduce new RuleExpr
types and forget to change it here.

Instead, the output type information is produced right where we do the
compilation, in RuleExprCompiler::process(). We can now look at the
output type and decide on the correct strategy to wrap the output in a
flattenable struct. No guessing required.

This commit ends the whole series for fixing the handling of optional
string components.