Fix constant optionals
We had a bug in cases like this, where optionals did not encompass any
parsed field:

    rule dbl_quoted_greeting: "hello [world]"

createOptionalRule() used to always assume that the part in it produces
a JsonLoc::Map. One that is always flattenable. But that is not always
true. The file `contact-line.oalex` shows another use case, where we
simply wanted an optional comma (`[,]`). This is now possible, while in
the past it used to crash with a Bug() in codegen.

This commit uses the new RuleOutputType that we had been piping through
the compilation layers for the last few commits. This way, we can tell
what output type to use in the fallback path of optional components not
matching.

The new implementation of createOptionalRule() is now constrained to
RuleOutputType::{string,flat_map}, just because that's what all current
callers need right now. We do not have a use case for expanding it right
now.