Move functions into anonymous namespace
The changes are to the `create*` functions. Also added a comment block
explaining the (lack of) convention used in this file. We will soon
change the return type of createOptionalRule() to be a new locally
defined type. The only way to not create tiny 5-line scopes of anonymous
blocks is to move these functions all into a large anonymous scope.