Make optional fields indirect
Previously, std::optional<FieldType> was used as the type for any
optional fields parsed. However, this does not work for recursive types
or incomplete types. This commit changes it to a new type,
oalex::indopt<FieldType> (which happens to be just an alias of
sz::any_of under the hood). It works mostly like std::optional, except
it uses a heap-allocated pointer internally, so that it is possible to
use fields of incomplete type.