Finish migrating to LoopRuleNewFields
This commit deletes the old LoopRuleFields, and renames the
LoopRuleNewFields struct to LoopRuleFields. All the constructor code is
migrated, so we no longer need a conversion glue (the conversion
operator).

This is a part of my push to make codegen "dumber". I want the input to
codegen to reflect the generated code more closely, so that:

  1. Codegen is simpler, has fewer bugs, and are easier to test.
  2. I can visualize the output code by just looking at the Rule being
     generated.

This change removes all notions of "part", "glue", "skip" etc. from
codegen. The various combinations of their values were getting harder
for me to follow, and it was becoming increasingly harder for me to
ensure correctness.