Move resolveIfWrapper() to analysis.cpp
Most callers don't really need to do a recursive "resolution" of types
any more. Instead, they can just read off the `typeSource` method of
`Rule`.

We only caller we needed to leave it as is in analysis.cpp because that
caller needs an integer ruleidx, while `typeSource` machinery provides a
more ephemeral `Rule&`.