Fix compRead fields for discarded components
It also sanity-checks the compRead() values produced by the compiler.

The compiler is sometimes sloppy, in that it sometimes assigns
unpackStruct to empty flat structs. This analysis phase fixes it.

For now, these computed compRead fields are not used anywhere, and are
recomputed by codegen as needed. Now that the values are fixed in this
commit codegen will soon be able to stop this recomputation.