Fix codegen for local structs
I don't have test harness for this yet, but here's how we see the
problem. First, run `oalex build` on an oalex file with this:

    rule asgn:
      var name = value
    where:
      name ~ /[a-z_]+/
      value ~ (v ~ /[0-9]+/)
    outputs: {
      name, value
    }

Now look at the `struct ParsedAsgn` created in the generated header. If
it creates a field with type name like `ParsedRule18 value`, that's a
problem. The type name has no stability for the user. On the other hand,
once this fix is done, it will output this:

    struct ParsedAsgn {
      oalex::LocPair loc;
      struct Fields {
        oalex::StringLoc name;
        using Value = ParsedRule18;
        Value value;
      } fields;
      explicit operator oalex::JsonLoc() const;
    };

Now the user can use `Value` as a stable type name.